combined the player with the best of what we got so far(in Adam_wallWalkTest-scene)
now the player is able to climb on any surface, in any angle, with the procedural animation and a functional camera too! Though the legs in the procedural animation does not take a higher speed, think it's set to 2 now, and by 5 it's already too fast.. this needs to be fixed further up.