added layerMask for raycasts, added "hook" raycast for climbing over edge
added new layer for PickUpAbles like tomatoes, etc, so that the raycast can ignore those normals when looking to climb.
Have an issue with climbing from under or over to the opposit side over a thinner edge. Possible solution is to move the backwards rays' origin more back and add more rays for down inbetween.