Fixed webswing-bug
the bug caused by holding UseWeb-button (left-click(Mouse) or Right-Trigger(Gamepad)) after shooting and then while still holding that button down, pressing the Aim-button (right-click(Mouse) or Left-Trigger(Gamepad)). This causes another attachment-point to be created and the first to remain, thus there are two and the player is stuck.